+++ mountpoint-s3-crt/src/auth.rs
@@ -21,5 +21,6 @@ fn auth_library_init(allocator: &Allocator) {
         unsafe {
             aws_auth_library_init(allocator.inner.as_ptr());
         }
+        crate::register_crt_cleanup_at_exit();
     });
 }