+++ mountpoint-s3-crt/src/common.rs
@@ -25,5 +25,6 @@ fn common_library_init(allocator: &Allocator) {
         unsafe {
             aws_common_library_init(allocator.inner.as_ptr());
         }
+        crate::register_crt_cleanup_at_exit();
     });
 }