+++ mountpoint-s3-client/CHANGELOG.md
@@ -1,5 +1,16 @@
-## Unreleased (v0.20.1)
+## Unreleased (v0.21.0)
 
+## v0.21.0 (July 20, 2026)
+
+### Breaking changes
+
+* `ObjectClient::put_object_single` now requires `contents: impl AsRef<[u8]> + Send + 'static` (was `+ 'a`), so the body can be held until the request is fully torn down. ([#1882](https://github.com/awslabs/mountpoint-s3/pull/1882))
+
+### Other changes
+
+* Add `TlsConfig` type and `S3ClientConfig::tls_config()` builder for configuring a custom CA trust store. ([#1834](https://github.com/awslabs/mountpoint-s3/pull/1834))
+* Run cleanup for CRT libraries on process exit. ([#1850](https://github.com/awslabs/mountpoint-s3/pull/1850))
+* Fix credential resolution when using a source profile with STS Web Identity. ([#1889](https://github.com/awslabs/mountpoint-s3/pull/1889))
 * Update to latest CRT dependencies.
 
 ## v0.20.0 (April 28, 2026)